+++ Changelog.md
@@ -2,6 +2,12 @@
 
 
 
+### 2.0.0 ( 2022-10-16)
+ - [x] Method 'publish' returns a promise;
+ - [x] Testing setup ( Cypress);
+
+
+
 ### 1.0.0 (2022-08-14)
  - [x] Initial code;
  - [x] Documentation;